requirements.txt
docker
gitpython
pandas
pytest
